Refactoring

+++ src/left_pane.cpp
@@ -3,12 +3,12 @@
 #include <algorithm>
 #include <math.h>
 
-left_pane::left_pane(int &curs_y, bool show_hidden_files, const std::string &search_str, 
+left_pane::left_pane(unsigned int &curs_y, bool show_hidden_files, const std::string &search_str, 
 		const std::string &current_path, unsigned int &page, unsigned int term_height) {
 	files = file_io::get_dir_files(current_path, show_hidden_files, search_str);
 	std::sort(files.begin(), files.end());
 	size = files.size();
-	const unsigned int current_line_number = (page - term_height) + static_cast<unsigned>(curs_y) + 1;
+	const unsigned int current_line_number = (page - term_height) + curs_y + 1;
 	if (size < current_line_number) { //if cursor was over a file and it was deleted
 		page = static_cast<unsigned int>(std::ceil(static_cast<double>(size)
 					/ static_cast<double>(term_height))) * term_height;
